requirements.txt
streamlit
python-dotenv
# local package 
-e .
